move VectorSupplier and CloseableReranker into JVectorLuceneOnDiskGraph
